Switch Image arg from (Int,Int) to (Double,Double)

+++ Image.hs
@@ -2,7 +2,7 @@ module Image where
 
 import Color
 
-type Image = (Int, Int) -> Color
+type Image = (Double, Double) -> Color
 
 alwaysBlue :: Image
 alwaysBlue = \(x,y) -> Blue
